Made the home page refresh every second
Added an additional thread for the homepage while allowing the main app to run on a separate thread, so the app can still take inputs while it's busy refreshing the page. It would be better to run it on 2 separate processes though , as the current solution still results in there being a slight delay and some distortion in the transition from the home screen to the advanced screen. But implementing separate processes seems to result in an error, or maybe  not, will try again later...

Regardless, the implementation of the threads helps to make the code a little more responsive, IMO.

Changed the ordering for the imports as well